+++ __tests__/language-server/semantic-token.test.ts
@@ -5,7 +5,7 @@ import {
   buildInitialized,
   buildTextDocumentDidOpen,
   buildTextDocumentSemanticTokenFull
-} from "./util/lsp";
+} from "../util/lsp";
 
 let server: LanguageServer | undefined;
 
@@ -29,66 +29,6 @@ afterAll(() => {
   if (server) {
     return server.stop();
   }
-});
-
-test("initialize", async done => {
-  const server = LanguageServer.spawn();
-
-  const request = buildInitialize();
-  const response = await server.sendRequest(request);
-
-  expect(response).toEqual({
-    jsonrpc: "2.0",
-    id: request.id,
-    result: {
-      capabilities: {
-        semanticTokensProvider: {
-          full: true,
-          legend: {
-            tokenModifiers: [
-              "declaration",
-              "definition",
-              "readonly",
-              "static",
-              "deprecated",
-              "abstract",
-              "async",
-              "modification",
-              "documentation",
-              "defaultLibrary"
-            ],
-            tokenTypes: [
-              "keyword",
-              "variable",
-              "string",
-              "number",
-              "operator",
-              "comment",
-              "function",
-              "struct",
-              "function",
-              "parameter",
-              "property"
-            ]
-          }
-        },
-        textDocumentSync: 2
-      },
-      serverInfo: { name: "nico-ls", version: "0.0.1" }
-    },
-    error: null
-  });
-
-  // initialized
-  {
-    const notification = buildInitialized();
-    await server.sendNotification(notification);
-  }
-
-  expect(server.isStopped()).toBeFalsy();
-
-  server.on("exit", done);
-  server.stop();
 });
 
 test("open a document", async done => {
